Rename components
Better reflect what they contain.
