+++ client/src/lib/Dashboard/AdvisoryQuery.svelte
