+++ client/src/lib/Dashboard/EventQuery.svelte
